+++ src/app/(main)/profile/mypage/layout.tsx
@@ -1,0 +1,9 @@
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  robots: { index: false, follow: false },
+};
+
+export default function MyPageLayout({ children }: { children: React.ReactNode }) {
+  return <>{children}</>;
+}
