+++ src/app/(public)/find-password/page.tsx
@@ -1,86 +1,10 @@
-"use client";
+import type { Metadata } from "next";
+import FindPasswordPageClient from "./FindPasswordPageClient";
 
-import React, { useState } from "react";
-import Image from "next/image";
-import { useRouter } from "next/navigation";
-import PrimaryButton from "@/components/common/find-account/PrimaryButton";
-import InputField from "@/components/common/find-account/InputField";
-import { toast } from "react-hot-toast";
-import { useSendTempPasswordMutation } from "@/hooks/mutations/useAuthMutations";
+export const metadata: Metadata = {
+  title: "비밀번호 찾기",
+};
 
 export default function FindPasswordPage() {
-    const router = useRouter();
-    const [email, setEmail] = useState("");
-
-    const { mutate: sendTempPassword, isPending } = useSendTempPasswordMutation();
-
-    const handleBack = () => {
-        router.back();
-    };
-
-    const handleSendTempPassword = () => {
-        if (!email) {
-            toast.error("이메일을 입력해주세요.");
-            return;
-        }
-
-        sendTempPassword(email, {
-            onSuccess: () => {
-                router.push("/");
-            }
-        });
-    };
-
-    return (
-        <div
-            className="min-h-screen w-full flex items-center justify-center bg-cover bg-center bg-no-repeat py-12"
-            style={{ backgroundImage: "url('/background.png')" }}
-        >
-            <div
-                className="flex w-[766px] px-[120px] py-[99px] items-center justify-center flex-col gap-[30px] rounded-lg bg-white shadow-xl"
-            >
-                {/* 상단 텍스트 및 로고 영역 */}
-                <div className="flex flex-col items-center gap-[80px] w-full">
-                    <div className="flex flex-col items-center gap-[24px]">
-                        <Image
-                            src="/Vector.svg"
-                            alt="CheckMo Logo"
-                            width={76}
-                            height={46}
-                            className="w-[76.607px] h-[46.815px] shrink-0"
-                        />
-                        <h1 className="text-primary-2 text-center subhead_1">
-                            비밀번호 재발급
-                        </h1>
-                    </div>
-                    <p className="text-Gray-6 text-center subhead_1">
-                        이메일을 입력해주세요
-                    </p>
-                </div>
-
-                <div className="flex flex-col items-center gap-[140px] w-full">
-                    {/* 폼 영역 */}
-                    <div className="flex flex-col items-start gap-[24px] w-full max-w-[526px]">
-                        <InputField
-                            type="email"
-                            placeholder="이메일"
-                            value={email}
-                            onChange={(e) => setEmail(e.target.value)}
-                            disabled={isPending}
-                        />
-                    </div>
-
-                    {/* 하단 버튼 그룹 */}
-                    <div className="flex items-center gap-[10px] justify-center w-full">
-                        <PrimaryButton onClick={handleBack} disabled={isPending}>
-                            뒤로가기
-                        </PrimaryButton>
-                        <PrimaryButton onClick={handleSendTempPassword} disabled={isPending}>
-                            {isPending ? "전송 중..." : "임시 비밀번호 전송"}
-                        </PrimaryButton>
-                    </div>
-                </div>
-            </div>
-        </div>
-    );
+  return <FindPasswordPageClient />;
 }
